# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "PageVetPetDTO" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/VetPetDTO"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "ascending" : {
            "type" : "boolean"
          },
          "direction" : {
            "type" : "string"
          },
          "ignoreCase" : {
            "type" : "boolean"
          },
          "nullHandling" : {
            "type" : "string"
          },
          "property" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "VaccineCorrectionResponseDTO" : {
        "properties" : {
          "correctedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "correctedByName" : {
            "type" : "string"
          },
          "correctedByOrganizationName" : {
            "type" : "string"
          },
          "previousApplicationDate" : {
            "format" : "date",
            "type" : "string"
          },
          "previousDescription" : {
            "type" : "string"
          },
          "previousNextDoseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "previousVaccineName" : {
            "type" : "string"
          },
          "vaccineCorrectionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "VaccineRequestDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "applicationDate" : {
            "format" : "date",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "nextDoseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "organizationId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "vaccineCatalogId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "vaccineName" : {
            "type" : "string"
          }
        },
        "required" : [ "animalId" ],
        "type" : "object"
      },
...
      "VaccineResponseDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "applicationDate" : {
            "format" : "date",
            "type" : "string"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "nextDoseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "organizationId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "updateDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "vaccineCatalogId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "vaccineId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "vaccineName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "VetPetDTO" : {
        "properties" : {
          "accessGrantedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "bornDate" : {
            "format" : "date",
            "type" : "string"
          },
          "breed" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "healthStatus" : {
            "enum" : [ "OVERDUE", "DUE_SOON", "UP_TO_DATE", "NO_NEXT_DOSE" ],
            "type" : "string"
          },
          "lastVisitAt" : {
            "format" : "date",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "personName" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "underTreatment" : {
            "type" : "boolean"
          },
          "weight" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "O token devolvido por POST /auth/login.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/professional/animals" : {
      "get" : {
        "operationId" : "listAccessibleAnimals",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "animal.name,ASC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageVetPetDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Os animais que o contexto ativo alcanca, com busca e ordem",
        "tags" : [ "professional-animal-controller" ]
      }
    },
    "/professional/animals/in-custody" : {
      "get" : {
        "operationId" : "listAnimalsInCustody",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "animal.name,ASC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageVetPetDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Os animais sob custodia da organizacao ativa",
        "tags" : [ "professional-animal-controller" ]
      }
    },
...
    "/professional/animals/{animalId}/vaccines" : {
      "get" : {
        "operationId" : "listVaccines",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/VaccineResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "As vacinas do animal, pelo lado profissional",
        "tags" : [ "professional-animal-controller" ]
      },
      "post" : {
        "description" : "A organizacao vem do contexto ativo e NUNCA do payload: aceitar organizationId do cliente deixaria um veterinario registrar em nome de outra clinica.",
        "operationId" : "registerVaccine",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VaccineRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VaccineResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra uma vacina no animal atendido",
        "tags" : [ "professional-animal-controller" ]
      }
    },
    "/professional/animals/{animalId}/vaccines/{vaccineId}" : {
      "put" : {
        "description" : "Correcao e nao reescrita: o estado anterior fica gravado. Vale so para registro do proprio contexto e so dentro de uma janela curta - corrigir lancamento de meses atras nao e conserto de digitacao, e o tutor e quem decide o que fica na carteira dele.",
        "operationId" : "correctVaccine",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "vaccineId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VaccineRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VaccineResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Corrige uma vacina que a propria organizacao registrou",
        "tags" : [ "professional-animal-controller" ]
      }
    },
    "/professional/animals/{animalId}/vaccines/{vaccineId}/corrections" : {
      "get" : {
        "operationId" : "listCorrections_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "vaccineId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/VaccineCorrectionResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "O historico de correcoes da vacina",
        "tags" : [ "professional-animal-controller" ]
      }
    },
...
    "/vaccines/{vaccineId}/corrections" : {
      "get" : {
        "description" : "Correcao e sucessao: o valor anterior fica legivel, com quem corrigiu e quando. A linha do tempo diz QUANTAS correcoes o evento sofreu; o conteudo delas esta aqui.",
        "operationId" : "listCorrections",
        "parameters" : [ {
          "in" : "path",
          "name" : "vaccineId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/VaccineCorrectionResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "O historico de correcoes da vacina",
        "tags" : [ "vaccine-controller" ]
      }
    }
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ]
}